Run Prettier on commit

+++ contracts/token/ERC1155/base/ERC1155BaseStorage.sol
@@ -3,17 +3,18 @@
 pragma solidity ^0.8.0;
 
 library ERC1155BaseStorage {
-  struct Layout {
-    mapping (uint => mapping (address => uint)) balances;
-    mapping (address => mapping (address => bool)) operatorApprovals;
-  }
+    struct Layout {
+        mapping(uint256 => mapping(address => uint256)) balances;
+        mapping(address => mapping(address => bool)) operatorApprovals;
+    }
 
-  bytes32 internal constant STORAGE_SLOT = keccak256(
-    'solidstate.contracts.storage.ERC1155Base'
-  );
+    bytes32 internal constant STORAGE_SLOT =
+        keccak256('solidstate.contracts.storage.ERC1155Base');
 
-  function layout () internal pure returns (Layout storage l) {
-    bytes32 slot = STORAGE_SLOT;
-    assembly { l.slot := slot }
-  }
+    function layout() internal pure returns (Layout storage l) {
+        bytes32 slot = STORAGE_SLOT;
+        assembly {
+            l.slot := slot
+        }
+    }
 }
