+++ contracts/token/ERC1155/enumerable/ERC1155EnumerableStorage.sol
@@ -2,21 +2,22 @@
 
 pragma solidity ^0.8.0;
 
-import {EnumerableSet} from '../../../utils/EnumerableSet.sol';
+import { EnumerableSet } from '../../../utils/EnumerableSet.sol';
 
 library ERC1155EnumerableStorage {
-  struct Layout {
-    mapping (uint => uint) totalSupply;
-    mapping (uint => EnumerableSet.AddressSet) accountsByToken;
-    mapping (address => EnumerableSet.UintSet) tokensByAccount;
-  }
+    struct Layout {
+        mapping(uint256 => uint256) totalSupply;
+        mapping(uint256 => EnumerableSet.AddressSet) accountsByToken;
+        mapping(address => EnumerableSet.UintSet) tokensByAccount;
+    }
 
-  bytes32 internal constant STORAGE_SLOT = keccak256(
-    'solidstate.contracts.storage.ERC1155Enumerable'
-  );
+    bytes32 internal constant STORAGE_SLOT =
+        keccak256('solidstate.contracts.storage.ERC1155Enumerable');
 
-  function layout () internal pure returns (Layout storage l) {
-    bytes32 slot = STORAGE_SLOT;
-    assembly { l.slot := slot }
-  }
+    function layout() internal pure returns (Layout storage l) {
+        bytes32 slot = STORAGE_SLOT;
+        assembly {
+            l.slot := slot
+        }
+    }
 }
