+++ contracts/token/ERC1155/enumerable/IERC1155Enumerable.sol
@@ -11,26 +11,32 @@ interface IERC1155Enumerable {
      * @param id token id to query
      * @return token supply
      */
-    function totalSupply (uint id) external view returns (uint);
+    function totalSupply(uint256 id) external view returns (uint256);
 
     /**
      * @notice query total number of holders for given token
      * @param id token id to query
      * @return quantity of holders
      */
-    function totalHolders (uint id) external view returns (uint);
+    function totalHolders(uint256 id) external view returns (uint256);
 
     /**
      * @notice query holders of given token
      * @param id token id to query
      * @return list of holder addresses
      */
-    function accountsByToken (uint id) external view returns (address[] memory);
+    function accountsByToken(uint256 id)
+        external
+        view
+        returns (address[] memory);
 
     /**
      * @notice query tokens held by given address
      * @param account address to query
      * @return list of token ids
      */
-    function tokensByAccount (address account) external view returns (uint[] memory);
+    function tokensByAccount(address account)
+        external
+        view
+        returns (uint256[] memory);
 }
