+++ contracts/token/ERC20/permit/ERC20PermitMock.sol
@@ -2,13 +2,13 @@
 
 pragma solidity ^0.8.0;
 
-import {ERC20MetadataStorage} from '../metadata/ERC20MetadataStorage.sol';
-import {ERC20Permit} from './ERC20Permit.sol';
+import { ERC20MetadataStorage } from '../metadata/ERC20MetadataStorage.sol';
+import { ERC20Permit } from './ERC20Permit.sol';
 
 contract ERC20PermitMock is ERC20Permit {
     using ERC20MetadataStorage for ERC20MetadataStorage.Layout;
 
-    constructor (
+    constructor(
         string memory name,
         string memory symbol,
         uint8 decimals
@@ -20,11 +20,11 @@ contract ERC20PermitMock is ERC20Permit {
         l.setDecimals(decimals);
     }
 
-    function __mint (address account, uint amount) external {
+    function __mint(address account, uint256 amount) external {
         _mint(account, amount);
     }
 
-    function __burn (address account, uint amount) external {
+    function __burn(address account, uint256 amount) external {
         _burn(account, amount);
     }
 }
