+++ contracts/token/ERC721/metadata/ERC721MetadataStorage.sol
@@ -3,19 +3,20 @@
 pragma solidity ^0.8.0;
 
 library ERC721MetadataStorage {
-  bytes32 internal constant STORAGE_SLOT = keccak256(
-    'solidstate.contracts.storage.ERC721Metadata'
-  );
+    bytes32 internal constant STORAGE_SLOT =
+        keccak256('solidstate.contracts.storage.ERC721Metadata');
 
-  struct Layout {
-    string name;
-    string symbol;
-    string baseURI;
-    mapping (uint => string) tokenURIs;
-  }
+    struct Layout {
+        string name;
+        string symbol;
+        string baseURI;
+        mapping(uint256 => string) tokenURIs;
+    }
 
-  function layout () internal pure returns (Layout storage l) {
-    bytes32 slot = STORAGE_SLOT;
-    assembly { l.slot := slot }
-  }
+    function layout() internal pure returns (Layout storage l) {
+        bytes32 slot = STORAGE_SLOT;
+        assembly {
+            l.slot := slot
+        }
+    }
 }
