+++ contracts/utils/MulticallMock.sol
@@ -5,11 +5,11 @@ pragma solidity ^0.8.0;
 import './Multicall.sol';
 
 contract MulticallMock is Multicall {
-  function callTest () external pure returns (uint) {
-    return 1;
-  }
+    function callTest() external pure returns (uint256) {
+        return 1;
+    }
 
-  function callRevertTest () external pure {
-    revert('revert');
-  }
+    function callRevertTest() external pure {
+        revert('revert');
+    }
 }
