+++ contracts/utils/ReentrancyGuardStorage.sol
@@ -3,16 +3,17 @@
 pragma solidity ^0.8.0;
 
 library ReentrancyGuardStorage {
-  struct Layout {
-    uint status;
-  }
+    struct Layout {
+        uint256 status;
+    }
 
-  bytes32 internal constant STORAGE_SLOT = keccak256(
-    'solidstate.contracts.storage.ReentrancyGuard'
-  );
+    bytes32 internal constant STORAGE_SLOT =
+        keccak256('solidstate.contracts.storage.ReentrancyGuard');
 
-  function layout () internal pure returns (Layout storage l) {
-    bytes32 slot = STORAGE_SLOT;
-    assembly { l.slot := slot }
-  }
+    function layout() internal pure returns (Layout storage l) {
+        bytes32 slot = STORAGE_SLOT;
+        assembly {
+            l.slot := slot
+        }
+    }
 }
